feat: email module

+++ tools/cleanupPackages.sh
@@ -2,8 +2,8 @@
 
 for dir in ./packages/*; do
   package_name=$(basename "$dir")
-  
-  if [ -d "./node_modules/@cmmv/$package_name" ]; then
+
+  if [ "$package_name" != "core" ] && [ -d "./node_modules/@cmmv/$package_name" ]; then
     echo "Removing ./node_modules/@cmmv/$package_name"
     rm -rf "./node_modules/@cmmv/$package_name"
   fi
